Moved resource path to Gruntfile.
Fonts and images from the style guide needs to be usable
in other projects. The current setup forces the images and
fonts to be located in "../fonts" and "../images" (I'm guessing
this is because the github path is knewton.github.io/styleguide/).
That does not work well with other projects. Moved path
configuration to Gruntfile so that any project can define their path.